+++ src/mock.rs
@@ -2,10 +2,10 @@
 
 use crate::action::Action;
 use crate::matcher::{Matcher, SoapRequest};
-use crate::responder::{ResponseBody, Responder};
+use crate::responder::{Responder, ResponseBody};
 use std::net::SocketAddr;
-use std::sync::atomic::{AtomicU32, Ordering};
 use std::sync::Arc;
+use std::sync::atomic::{AtomicU32, Ordering};
 use std::time::Instant;
 use tokio::sync::RwLock;